(#50) Move status_line inside of OutputBuffer
